added first name, last name and email to user class. Updated decode user to include admin email, used auth_service instead of generator
