+++ breeze/models/admin.py
@@ -2,28 +2,7 @@ from .user import User
 
 class Admin(User):
     def __init__(self, username, password, first_name = None, last_name = None, email= None, is_disabled=False):
-        super().__init__(username, password, role="Admin", is_disabled=is_disabled)
-        self.__first_name = first_name
-        self.__last_name = last_name
-        self.__email = email
-
-    def get_first_name(self):
-        return self.__first_name
-    
-    def set_first_name(self, first_name):
-        self.__first_name = first_name
-    
-    def get_last_name(self):
-        return self.__last_name
-    
-    def set_last_name(self, last_name):
-        self.__last_name = last_name    
-    
-    def get_email(self):
-        return self.__email
-    
-    def set_email(self, email):
-        self.__email = email
+        super().__init__(username, password, role="Admin", first_name= first_name, last_name= last_name, email= email, is_disabled=is_disabled)
     
     def __str__(self):
         return f"Admin: {self.get_username()}, Role: {self.get_role()}"
