+++ breeze/models/mhwp.py
@@ -2,29 +2,8 @@ from .user import User
 
 class MHWP(User):
     def __init__(self, username, password, first_name = None, last_name = None, email= None, is_disabled=False, appointments=[]):
-        super().__init__(username, password, role='MHWP', is_disabled=is_disabled)
-        self.__first_name = first_name
-        self.__last_name = last_name
-        self.__email = email
+        super().__init__(username, password, role='MHWP', first_name= first_name, last_name= last_name, email= email, is_disabled=is_disabled)
         self.__appointments = appointments
-    
-    def get_first_name(self):
-        return self.__first_name
-
-    def set_first_name(self, first_name):
-        self.__first_name = first_name
-
-    def get_last_name(self):
-        return self.__last_name
-
-    def set_last_name(self, last_name):
-        self.__last_name = last_name
-
-    def get_email(self):
-        return self.__email
-    
-    def set_email(self, email):
-        self.__email = email
 
     def __str__(self):
         return f"MHWP: {self.get_username()}, Role: {self.get_role()}"
