+++ breeze/utils/data_utils.py
@@ -37,6 +37,7 @@ def decode_user(user_data):
             password=user_data.get("password"),
             first_name=user_data.get("information", {}).get("firstName"),
             last_name=user_data.get("information", {}).get("lastName"),
+            email=user_data.get("information", {}).get("email"),
             is_disabled=user_data.get("isDisabled", False)
         )
